refactor organization skill manage rows
